Fix admitted lemma issue from last commit. The fact that there was an admitted lemma was a symptom of a bigger issue:
The concrete semantics was evaluating the match condition once on the original state *before* going through the list of header ops.

The symbolic semantics was evaluating the match condition repeatedly on the updated state as it was going through the list of header ops.

The previous lemma had to be admitted because there was no way to prove equivalence of symbolic and concrete semantics when they were modeling different behaviors of the match condition. This is a good example of the kind of bug that Rocq lets us catch.

In reality, the match condition should be evaluated once at the beginning against the original state like the concrete semantics was doing. This commit fixed the symbolic semantics to match with the concrete one, and now the proof goes through without any admits.